requirements.txt
# Dépendances du projet Carte42 - Détection de changement PCRS
# Géospatial
geopandas>=0.14.0
rasterio>=1.3.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0

# Traitement image
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Visualisation & cartographie
folium>=0.15.0
matplotlib>=3.7.0
branca>=0.7.0

# Réseau & données
requests>=2.31.0
tqdm>=4.66.0

# Utilitaires
python-dotenv>=1.0.0
